+++ demisto_sdk/commands/update_release_notes/update_rn.py
@@ -11,10 +11,12 @@ from typing import Any, Optional, Tuple, Union
 
 from demisto_sdk.commands.common.constants import (
     ALL_FILES_VALIDATION_IGNORE_WHITELIST, DEFAULT_ID_SET_PATH,
-    IGNORED_PACK_NAMES, RN_HEADER_BY_FILE_TYPE, FileType)
+    DEPRECATED_REGEXES, IGNORED_PACK_NAMES, RN_HEADER_BY_FILE_TYPE, FileType)
 from demisto_sdk.commands.common.content import Content
 from demisto_sdk.commands.common.content.objects.pack_objects import (
     Integration, Playbook, Script)
+from demisto_sdk.commands.common.content.objects.pack_objects.abstract_pack_objects.yaml_content_object import \
+    YAMLContentObject
 from demisto_sdk.commands.common.git_util import GitUtil
 from demisto_sdk.commands.common.handlers import JSON_Handler
 from demisto_sdk.commands.common.tools import (LOG_COLORS, find_type,
@@ -34,17 +36,28 @@ json = JSON_Handler()
 CLASS_BY_FILE_TYPE = {FileType.INTEGRATION: Integration, FileType.SCRIPT: Script, FileType.PLAYBOOK: Playbook}
 
 
-def commands_to_deprecate_set(commands: list) -> set:
+def get_deprecated_comment_from_desc(description: str) -> str:
+    """
+    find deprecated comment from description
+    Args:
+        description: The yml description
+
+    Returns:
+        If a deprecated description is found return it for rn.
+    """
+    deprecate_line_with_replacement = re.findall(DEPRECATED_REGEXES[0], description)
+    deprecate_line_no_replacement = re.findall(DEPRECATED_REGEXES[1], description)
+
+    deprecate_line = deprecate_line_with_replacement + deprecate_line_no_replacement
+    return deprecate_line[0] if deprecate_line else ''
+
+
+def deprecated_commands(commands: list) -> set:
     """return a set of the deprecated commands only"""
-    res_set = set()
-    for command in commands:
-        if command.get('deprecated'):
-            res_set.add(command.get('name'))
-
-    return res_set
+    return {command.get('name') for command in commands if command.get('deprecated')}
 
 
-def get_yml_objects(path: str, file_type) -> Tuple[Any, Union[Integration, Script, Playbook]]:
+def get_yml_objects(path: str, file_type) -> Tuple[Any, YAMLContentObject]:
     """
     Generate yml files from master and from the current branch
     Args:
@@ -65,15 +78,17 @@ def get_deprecated_rn(path: str, file_type):
     old_yml, new_yml = get_yml_objects(path, file_type)
 
     if not old_yml.get('deprecated') and new_yml.is_deprecated:
-        return '- Deprecated. Use %%% instead.\n'
+        description = new_yml.get('comment') if file_type == file_type.SCRIPT else new_yml.get('description')
+        rn_from_description = get_deprecated_comment_from_desc(description)
+        return f'- Deprecated. {rn_from_description or "Use %%% instead"}.\n'
 
     if file_type != FileType.INTEGRATION:
         return ''
 
     # look for deprecated commands
     rn = ''
-    old_commands = commands_to_deprecate_set(old_yml.get('script', {}).get('commands'))
-    new_commands = commands_to_deprecate_set(new_yml.script.get('commands'))
+    old_commands = deprecated_commands(old_yml.get('script', {}).get('commands'))
+    new_commands = deprecated_commands(new_yml.script.get('commands'))
 
     for command_name in new_commands:
         # if command is deprecated in new yml, and not in old yml
@@ -604,7 +619,9 @@ class UpdateRN:
                     if _type in (FileType.INTEGRATION, FileType.SCRIPT, FileType.PLAYBOOK):
                         deprecate_rn = get_deprecated_rn(path, _type)
                     if deprecate_rn:
-                        rn_desc += f'- {text}\n{deprecate_rn}'
+                        if text:
+                            rn_desc += f'- {text}\n'
+                        rn_desc += deprecate_rn
                     else:
                         rn_desc += f'- {text or "%%UPDATE_RN%%"}\n'